Moved all getResource methods into test tree.
