+++ src/test/java/org/apache/datasketches/quantiles/QuantilesSketchCrossLanguageTest.java
@@ -24,7 +24,6 @@ import static org.apache.datasketches.common.TestUtil.CHECK_CPP_HISTORICAL_FILES
 import static org.apache.datasketches.common.TestUtil.GENERATE_JAVA_FILES;
 import static org.apache.datasketches.common.TestUtil.cppPath;
 import static org.apache.datasketches.common.TestUtil.javaPath;
-import static org.apache.datasketches.common.Util.getResourceBytes;
 import static org.apache.datasketches.quantilescommon.QuantileSearchCriteria.EXCLUSIVE;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
@@ -34,6 +33,7 @@ import java.nio.file.Files;
 import java.util.Comparator;
 
 import org.apache.datasketches.common.ArrayOfStringsSerDe;
+import org.apache.datasketches.common.TestUtil;
 import org.apache.datasketches.memory.Memory;
 import org.apache.datasketches.quantilescommon.QuantilesDoublesSketchIterator;
 import org.apache.datasketches.quantilescommon.QuantilesGenericSketchIterator;
@@ -238,7 +238,7 @@ public class QuantilesSketchCrossLanguageTest {
     println("fullName: "+ fileName);
     println("Old Median: " + quantile);
     //Read File bytes
-    byte[] byteArr = getResourceBytes(fileName);
+    byte[] byteArr = TestUtil.getResourceBytes(fileName);
     Memory srcMem = Memory.wrap(byteArr);
 
     // heapify as update sketch
